@@ -7,6 +7,12 @@ using System.Text;
 using System.Text.RegularExpressions;
 using Rotativa.AspNetCore.Options;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Mvc.ModelBinding;
+using Microsoft.AspNetCore.Mvc.Rendering;
+using Microsoft.AspNetCore.Mvc.ViewEngines;
+
+using Microsoft.AspNetCore.Mvc.ViewFeatures;
+
 #if NET5_0_OR_GREATER
     using Microsoft.AspNetCore.Mvc;
     using Microsoft.AspNetCore.Http;
@@ -20,13 +26,31 @@ using System.Threading.Tasks;
 
 namespace Rotativa.AspNetCore
 {
-    public abstract class AsResultBase : ViewResult //IActionResult
+    public abstract class AsResultBase : ViewResult
     {
         protected AsResultBase()
         {
             this.WkhtmlPath = string.Empty;
             this.FormsAuthenticationCookieName = ".ASPXAUTH";
+            this.IsPartialView = false;
+            this.SetBaseUrl = true;
+            this.DontStopSlowScripts = false;
+            this.NoImages = false;
         }
+
+        /// <summary>
+        /// Determines if the view that is referenced is partial or not.
+        /// </summary>
+        public bool IsPartialView { get; set; }
+
+        /// <summary>
+        /// Whether we add a base URL when we generate the HTML.
+        /// </summary>
+        /// <remarks>
+        /// This was always on because it wasn't configurable (<= 1.3.2). That's why the default is on.
+        /// However it's cleaner to allow developers to set it themselves, and only add the BaseUrl when requested.
+        /// </remarks>
+        public bool SetBaseUrl { get; set; }
 
         /// <summary>
         /// This will be send to the browser as a name of the generated PDF file.
@@ -34,7 +58,7 @@ namespace Rotativa.AspNetCore
         public string FileName { get; set; }
 
         /// <summary>
-        /// Path to wkhtmltopdf\wkhtmltoimage binary.
+        /// Path to wkhtmltopdf / wkhtmltoimage binary.
         /// </summary>
         public string WkhtmlPath { get; set; }
 
@@ -78,10 +102,34 @@ namespace Rotativa.AspNetCore
         public bool IsJavaScriptDisabled { get; set; }
 
         /// <summary>
+        /// Indicates whether the page can run JavaScript.
+        /// </summary>
+        [OptionFlag("--no-stop-slow-scripts")]
+        public bool DontStopSlowScripts { get; set; }
+
+        /// <summary>
+        /// Specify a user style sheet, to load with every page.
+        /// </summary>
+        [OptionFlag("--user-style-sheet")]
+        public string UserStyleSheet { get; set; }
+
+        /// <summary>
         /// Minimum font size.
         /// </summary>
         [OptionFlag("--minimum-font-size")]
         public int? MinimumFontSize { get; set; }
+
+        /// <summary>
+        /// Sets the zoom level.
+        /// </summary>
+        [OptionFlag("--zoom")]
+        public double? Zoom { get; set; }
+
+        /// <summary>
+        /// Do not load or print images.
+        /// </summary>
+        [OptionFlag("--no-images")]
+        public bool NoImages { get; set; }
 
         /// <summary>
         /// Sets proxy server.
@@ -102,6 +150,18 @@ namespace Rotativa.AspNetCore
         public string Password { get; set; }
 
         /// <summary>
+        /// Password to ssl client cert private key.
+        /// </summary>
+        [OptionFlag("--ssl-key-password")]
+        public string SslKeyPassword { get; set; }
+
+        /// <summary>
+        /// Path to the ssl client cert public key in OpenSSL PEM format, optionally followed by intermediate ca and trusted certs
+        /// </summary>
+        [OptionFlag("--ssl-crt-path")]
+        public string SslCrtPath { get; set; }
+
+        /// <summary>
         /// Use this if you need another switches that are not currently supported by Rotativa.
         /// </summary>
         [OptionFlag("")]
@@ -115,10 +175,10 @@ namespace Rotativa.AspNetCore
         protected abstract string GetUrl(Microsoft.AspNetCore.Mvc.ActionContext context);
 
         /// <summary>
-        /// Returns properties with OptionFlag attribute as one line that can be passed to wkhtmltopdf binary.
+        /// Returns properties with OptionFlag attribute as one line that can be passed to wkhtmltopdf / wkhtmltoimage binary.
         /// </summary>
-        /// <returns>Command line parameter that can be directly passed to wkhtmltopdf binary.</returns>
-        protected virtual string GetConvertOptions()
+        /// <returns>Command line parameter that can be directly passed to wkhtmltopdf / wkhtmltoimage binary.</returns>
+        public virtual string GetConvertOptions()
         {
             var result = new StringBuilder();
 
@@ -244,5 +304,72 @@ namespace Rotativa.AspNetCore
         }
 
         protected abstract string GetContentType();
+
+        /// <summary>
+        /// Get the view out of the context.
+        /// </summary>
+        /// <param name="context">The action context</param>
+        /// <returns></returns>
+        /// <exception cref="InvalidOperationException"></exception>
+        protected IView GetView(ActionContext context)
+        {
+            // Use current action name if the view name was not provided
+            if (string.IsNullOrEmpty(ViewName))
+            {
+                ViewName = ((Microsoft.AspNetCore.Mvc.Controllers.ControllerActionDescriptor)context.ActionDescriptor).ActionName;
+            }
+
+            var engine = context.HttpContext.RequestServices.GetService(typeof(ICompositeViewEngine)) as ICompositeViewEngine;
+            var getViewResult = engine.GetView(executingFilePath: null, viewPath: ViewName, isMainPage: !IsPartialView);
+            if (getViewResult.Success)
+            {
+                return getViewResult.View;
+            }
+
+            var findViewResult = engine.FindView(context, ViewName, isMainPage: !IsPartialView);
+            if (findViewResult.Success)
+            {
+                return findViewResult.View;
+            }
+
+            var searchedLocations = getViewResult.SearchedLocations.Concat(findViewResult.SearchedLocations);
+            var errorMessage = string.Join(
+                System.Environment.NewLine,
+                new[] { $"Unable to find view '{ViewName}'. The following locations were searched:" }.Concat(searchedLocations));
+
+            throw new InvalidOperationException(errorMessage);
+        }
+
+        protected async Task<string> GetHtmlFromView(ActionContext context)
+        {
+            var view = GetView(context);
+            var html = new StringBuilder();
+
+            ITempDataProvider tempDataProvider = context.HttpContext.RequestServices.GetService(typeof(ITempDataProvider)) as ITempDataProvider;
+            var tempDataDictionary = new TempDataDictionary(context.HttpContext, tempDataProvider);
+
+            using (var output = new StringWriter())
+            {
+                var viewContext = new ViewContext(
+                    context,
+                    view,
+                    this.ViewData,
+                    tempDataDictionary,
+                    output,
+                    new HtmlHelperOptions());
+
+                await view.RenderAsync(viewContext);
+
+                html = output.GetStringBuilder();
+            }
+
+            if (!this.SetBaseUrl)
+            {
+                return html.ToString();
+            }
+
+            string baseUrl = string.Format("{0}://{1}", context.HttpContext.Request.Scheme, context.HttpContext.Request.Host);
+            return Regex.Replace(html.ToString(), "<head>", string.Format("<head><base href=\"{0}\" />", baseUrl), RegexOptions.IgnoreCase);
+        }
     }
 }
